requirements.txt
gradio
torch~=1.2
torchvision 
opencv-python~=3.4
ftfy
regex
tqdm
pycrypto
numpy
Pillow